net: widen entity_state effects to uint32_t; bump PROTOCOL_MINOR to 1025
EF_SHADOWS (bit 16) and EF_PENTAGRAM (bit 17) both overflowed uint16_t,
causing the flags to be silently dropped on the wire. Switch Net_WriteShort/
Net_ReadShort to Net_WriteLong/Net_ReadLong for the effects field.

+++ src/shared/shared.h
@@ -203,7 +203,7 @@ typedef struct {
   /**
    * @brief Bit mask of active visual effects (EF_ROTATE, EF_BOB, etc.).
    */
-  uint16_t effects;
+  uint32_t effects;
 
   /**
    * @brief Trail effect identifier (`entity_trail_t`).
